Final gradients still dont match. Maybe error in c matrix?

+++ Code/OilProblem/minimizer.cpp
@@ -79,7 +79,7 @@ struct PermeabilitiesProblem: public Problem<Real> {
             // row-major). It will give you the same hash value for two different matrices if they
             // are the transpose of each other in different storage order.
             size_t seed = 0;
-            for (size_t i = 0; i < matrix.size(); ++i) {
+            for (int i = 0; i < matrix.size(); ++i) {
                 auto elem = *(matrix.data() + i);
                 seed ^= std::hash<typename T::Scalar>()(elem) + 0x9e3779b9 + (seed << 6) + (seed >> 2);
             }
